+++ rkyv/src/impls/alloc/with/niching.rs
@@ -46,69 +46,35 @@ where
 
 #[cfg(test)]
 mod tests {
-    use core::num::NonZeroU32;
 
     use crate::{
-        api::test::roundtrip_with,
-        niche::niching::{DefaultNicher, NaN, Zero},
-        with::Nicher,
-        Archive, Deserialize, Serialize,
+        api::test::roundtrip_with, niche::niching::DefaultNicher, Archive,
+        Deserialize, Serialize,
     };
 
     #[derive(Archive, Serialize, Deserialize, Debug, PartialEq)]
     #[rkyv(crate, derive(Debug))]
     struct Nichable {
-        #[rkyv(niche = NaN)]
-        not_nan: f32,
-        #[rkyv(niche = Zero)]
-        int: NonZeroU32,
         #[rkyv(niche)] // Default = Null
         boxed: Box<i32>,
-    }
-
-    impl Nichable {
-        fn create() -> Self {
-            Nichable {
-                not_nan: 123.456,
-                int: unsafe { NonZeroU32::new_unchecked(789) },
-                boxed: Box::new(727),
-            }
-        }
     }
 
     #[test]
     fn with_struct() {
         #[derive(Archive, Serialize, Deserialize, Debug, PartialEq)]
         #[rkyv(crate, derive(Debug))]
-        struct Middle {
-            #[rkyv(with = Nicher<Zero>, niche = NaN, niche)] // Default = Null
-            a: Option<Nichable>,
-            #[rkyv(with = Nicher<NaN>, niche = Zero)]
-            b: Option<Nichable>,
-        }
-
-        #[derive(Archive, Serialize, Deserialize, Debug, PartialEq)]
-        #[rkyv(crate, derive(Debug))]
         struct Outer {
             #[rkyv(with = DefaultNicher)]
-            field: Option<Middle>,
+            field: Option<Nichable>,
         }
 
-        assert_eq!(
-            size_of::<ArchivedMiddle>(),
-            2 * size_of::<ArchivedNichable>()
-        );
-        assert_eq!(size_of::<ArchivedOuter>(), size_of::<ArchivedMiddle>());
+        assert_eq!(size_of::<ArchivedNichable>(), size_of::<ArchivedOuter>());
 
         let values = [
             Outer { field: None },
             Outer {
-                field: Some(Middle { a: None, b: None }),
-            },
-            Outer {
-                field: Some(Middle {
-                    a: None,
-                    b: Some(Nichable::create()),
+                field: Some(Nichable {
+                    boxed: Box::new(727),
                 }),
             },
         ];
@@ -117,17 +83,8 @@ mod tests {
             assert!(archived.field.is_none());
         });
         roundtrip_with(&values[1], |_, archived| {
-            let middle = archived.field.as_ref().unwrap();
-            assert!(middle.a.is_none());
-            assert!(middle.b.is_none());
-        });
-        roundtrip_with(&values[2], |_, archived| {
-            let middle = archived.field.as_ref().unwrap();
-            assert!(middle.a.is_none());
-            let b = middle.b.as_ref().unwrap();
-            assert_eq!(b.not_nan, 123.456);
-            assert_eq!(b.int.get(), 789);
-            assert_eq!(*b.boxed.as_ref(), 727);
+            let nichable = archived.field.as_ref().unwrap();
+            assert_eq!(nichable.boxed.as_ref().to_native(), 727);
         });
     }
 }
